+++ modules/core/04-channel/keeper/handshake.go
@@ -105,17 +105,13 @@ func (k Keeper) ChanOpenTry(
 	proofInit []byte,
 	proofHeight exported.Height,
 ) (string, *capabilitytypes.Capability, error) {
-	var (
-		channelID string
-	)
-
 	// connection hops only supports a single connection
 	if len(connectionHops) != 1 {
 		return "", nil, sdkerrors.Wrapf(types.ErrTooManyConnectionHops, "expected 1, got %d", len(connectionHops))
 	}
 
 	// generate a new channel
-	channelID = k.GenerateChannelIdentifier(ctx)
+	channelID := k.GenerateChannelIdentifier(ctx)
 
 	if !k.portKeeper.Authenticate(ctx, portCap, portID) {
 		return "", nil, sdkerrors.Wrapf(porttypes.ErrInvalidPort, "caller does not own port capability for port ID %s", portID)